Enable use of StreamResponse in `file/{fileId}/download` with a config flag
By setting a boolean flag 'music.experimental.streaming' in the config.php
file, it is now possible to enable the use of StreamResponse instead of
FileResponse to load the audio files on the web UI. This could possible speed
up the start of the playback on some systems using external storage, but this
has not been tested. It does seem to cause some problems with seeking, and
every now and then, there comes an error from the HTML5 audio element which
causes the playback to jump to the next track.