Small optimization on listing audio files from the library path
When getting all the audio files belonging to the user's library path, the
function LibrarySettings::pathBelongsToMusicLibrary gets called for each
audio file. The class LibrarySettings now caches some of the intermediate
values needed there which are fetched from the API of the cloud server. With
tens of thousands of files, this makes a measurable difference. With 60k
audio files, the measured time saving was around 2 seconds.